added separate editNotifications mutation
